==> project-2/Code/part5_test_data_cleaning.ipynb <==
# Project 2: Features, Galore: A Linear Regression Analysis on Home Features Predicting Sale Price
---

## Test Data Cleaning and Pre-processing

This section includes test data that will be ran through the optimal Linear Regression model. The data is cleaned and pre-processed with the same steps as the training data. The following steps are outlined below. 

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.preprocessing import StandardScaler

#Reading in the test.csv data 
data=pd.read_csv('../datasets/test.csv')
X_train=pd.read_csv('../datasets/X_train_pre_ss.csv')

data.head()

data.info()

### Converting column names to lowercase snake case

#Function from Ben's brain to convert column names to snakecase 
import re 

def snake_case(df):
    new_cols = {col: re.sub(r"([a-z]{1})([A-Z]{1})", r"\1 \2", col).replace(" ", "_").lower() for col in df.columns}
    return df.rename(columns = new_cols, inplace = True)

#Applying function to dataframe 
snake_case(data)

data.head()

#### Initial Data Cleaning 

#Making data ms_sub_class an object
data['ms_sub_class']=data['ms_sub_class'].astype('str')

#Dropping rows with minimal null values 
(data.dropna(subset=['bsmt_fin_sf_1', 'bsmt_fin_sf_2','bsmt_unf_sf', 'total_bsmt_sf',
                     'garage_cars', 'garage_area'], inplace=True))

#Changing central air to binary b/c missing no values (1 and O)
data['central_air']=np.where(data['central_air']=='Y', 1, 0)

data.info()

#### Feature Engineering New Columns/Dropping redundancies  

#Making new columns for variables that should be yes or no (ie. 1 for yes and 0 for no). No values need to be filled so okay now. 
data['has_wood_deck']=np.where(data['wood_deck_sf'] != 0, 1,0)
data['has_open_porch']=np.where(data['open_porch_sf'] != 0, 1,0)
data['has_enclosed_porch']=np.where(data['enclosed_porch']!= 0, 1,0)
data['has_3season_porch']=np.where(data['3ssn_porch'] != 0, 1, 0)
data['has_screen_porch']=np.where(data['screen_porch'] != 0, 1, 0)
data['has_pool']=np.where(data['pool_area'] != 0, 1, 0)

#Dropping old columns from dataframe 
data.drop(columns=['wood_deck_sf', 'open_porch_sf', 'enclosed_porch', '3ssn_porch', 'screen_porch', 'pool_area'], inplace=True)

#Dropping 'garage_area' due to redundancy with 'garage_cars'
data.drop(columns=['garage_area'], inplace=True)

#Dropping because redundant with garage_qual 
data.drop(columns=['garage_cond'], inplace=True)

#Dropping because likely correlated with the mas_vnr_type and I think mas_vnr_type is more important
data.drop(columns=['mas_vnr_area'], inplace=True)

data.head()

#### Replacing numerical null values with median 

#Replace null values with median in 'Lot Frontage'
data['lot_frontage']=data['lot_frontage'].replace(np.nan, data['lot_frontage'].median())

#Replace null values with median in 'Bsmt Full Bath'
data['bsmt_full_bath']=data['bsmt_full_bath'].replace(np.nan, data['bsmt_full_bath'].median())

#Replace null values with median in 'Bsmt Half Bath'
data['bsmt_half_bath']=data['bsmt_half_bath'].replace(np.nan, data['bsmt_half_bath'].median())

#Replace null values with median in 'Garage Yr Blt'
data['garage_yr_blt']=data['garage_yr_blt'].replace(np.nan, data['garage_yr_blt'].median())

#### Replacing categorical null values with 0

#Filling categorical columns with 0
data['alley']=data['alley'].replace(np.nan, 0)
data['mas_vnr_type']=data['mas_vnr_type'].replace(np.nan, 0)
data['bsmt_qual']=data['bsmt_qual'].replace(np.nan, 0)
data['bsmt_cond']=data['bsmt_cond'].replace(np.nan, 0)
data['bsmt_exposure']=data['bsmt_exposure'].replace(np.nan, 0)
data['bsmt_fin_type_1']=data['bsmt_fin_type_1'].replace(np.nan, 0)
data['bsmt_fin_type_2']=data['bsmt_fin_type_2'].replace(np.nan, 0)
data['fireplace_qu']=data['fireplace_qu'].replace(np.nan, 0)
data['garage_type']=data['garage_type'].replace(np.nan, 0)
data['garage_finish']=data['garage_finish'].replace(np.nan, 0)
data['garage_qual']=data['garage_qual'].replace(np.nan, 0)
data['pool_qc']=data['pool_qc'].replace(np.nan, 0)
data['fence']=data['fence'].replace(np.nan, 0)
data['misc_feature']=data['misc_feature'].replace(np.nan, 0)

data.info()

#### Ordering Ordinal Columns 

#Handeling ordinal columns

ordinal_dict = [
#Lot Shape
{'IR3': 1, 'IR2': 2, 'IR1': 3, 'Reg': 4},
    
#Utilities
{'AllPub':4,'NoSewr':3,'NoSeWa':2,'ELO':1},

#Land Slope
{'Gtl':3,'Mod':2,'Sev':1},
    
#Exter Qual
{'Ex':5,'Gd':4, 'TA':3, 'Fa':2,'Po':1}, 

#Exter Cond
{'Ex':5,'Gd':4, 'TA':3, 'Fa':2,'Po':1},

#Bsmt Qual
{'Ex':5,'Gd':4,'TA':3,'Fa':2,'Po':1, 0:0},
    
#Bsmt Cond
{'Ex':5,'Gd':4,'TA':3,'Fa':2,'Po':1, 0:0},
    
#Bsmt Exposure
{'Gd':4,'Av':3,'Mn':2,'No':1, 0:0},

#BsmtFin Type 1
{'GLQ':6,'ALQ':5,'BLQ':4,'Rec':3,'LwQ':2,'Unf':1, 0:0},

#BsmtFinType 2
{'GLQ':6,'ALQ':5,'BLQ':4,'Rec':3,'LwQ':2,'Unf':1, 0:0},

#HeatingQC 
{'Ex':5,'Gd':4,'TA':3,'Fa':2,'Po':1},

#Electrical
{'SBrkr':5,'FuseA':4,'FuseF':3,'FuseP':2,'Mix':1},

#KitchenQual
{'Ex':5,'Gd':4,'TA':3,'Fa':2,'Po':1},

#Functional
{'Typ':8,'Min1':7,'Min2':6,'Mod':5,'Maj1':4,'Maj2':3,'Sev':2,'Sal':1},

#FireplaceQu
{'Ex':5,'Gd':4,'TA':3,'Fa':2,'Po':1, 0:0},

#Garage Finish
{'Fin':3,'RFn':2,'Unf':1, 0:0},

#Garage Qual
{'Ex':5,'Gd':4,'TA':3,'Fa':2,'Po':1, 0:0},

#Paved Drive
{'Y':3,'P':2,'N':1},

#Pool QC
{'Ex':5,'Gd':4,'TA':3,'Fa':2,'Po':1, 0:0},

#Fence
{'GdPrv':4,'MnPrv':3,'GdWo':2,'MnWw':1, 0:0}
]

ordinal_columns = ['lot_shape', 'utilities', 'land_slope', 'exter_qual', 'exter_cond', 'bsmt_qual',
       'bsmt_cond', 'bsmt_exposure', 'bsmt_fin_type_1', 'bsmt_fin_type_2',
       'heating_qc', 'electrical', 'kitchen_qual', 'functional',
       'fireplace_qu', 'garage_finish', 'garage_qual',
       'paved_drive', 'pool_qc', 'fence']

#zip column names to dictionary values
ordinal_dicts = dict(zip(ordinal_columns ,ordinal_dict))

def ordinal_replace(df,columns):
    for column in columns:
        df[column] = df[column].map(ordinal_dicts[column])
    return df

#Calling function on training data
data=ordinal_replace(data, ordinal_columns)
data.head()

#### One-Hot-Encoding Categorical Columns with "Get_dummies"

#Dummy test.csv nominal columns 
data=pd.get_dummies(data, columns=['ms_sub_class','ms_zoning',
                                         'street', 'alley', 'land_contour', 'lot_config', 
                                         'neighborhood', 'condition_1', 'condition_2', 
                                         'bldg_type', 'house_style', 'roof_style', 'roof_matl', 
                                         'exterior_1st', 'exterior_2nd', 'mas_vnr_type', 
                                         'foundation', 'heating', 'garage_type', 'misc_feature', 
                                         'sale_type'], drop_first=True)

#### Bringing in X_train Un-scaled Dataset 

This X_train dataset is what the previous models were built upon. This dataset has already been cleaned, values imputed, ordinal columns handled and dummy columns created. This dataset will now be scaled with the training data and all other pre-modeling steps will be repeated to X_train and the test data accordingly. 


X_train.head()

X_train.rename(columns={"Unnamed: 0": "index"}, inplace=True)

X_train=X_train.set_index('index')

X_train.head()

#### Dropping all columns that are not in both datasets from X_train and test data 

#Getting datasets to match 
set(X_train.columns)-set(data.columns)

#Dropping columns in X_train that are not in test data

X_train.drop(columns=['condition_2_Feedr','condition_2_PosN','condition_2_RRAe','condition_2_RRNn',
                      'exterior_1st_CBlock','exterior_1st_ImStucc','exterior_1st_Stone','exterior_2nd_Stone',
                      'heating_OthW','heating_Wall','misc_feature_TenC','ms_sub_class_150','ms_zoning_C (all)',
                      'neighborhood_GrnHill','neighborhood_Landmrk','roof_matl_Membran'], inplace=True)

#Getting datasets to match 
set(data.columns)-set(X_train.columns)

#Dropping columns in test data that are not in X_train

data.drop(columns=['exterior_1st_AsphShn','exterior_1st_PreCast','exterior_2nd_Other','exterior_2nd_PreCast',
                   'heating_GasA','mas_vnr_type_CBlock','ms_zoning_I (all)','roof_matl_Metal','roof_matl_Roll',
                   'sale_type_VWD'], inplace=True)

X_train.shape

data.shape

#### Scaling the data using StandardScaler Class 

#Scaling data, dropping ID and PID so that they don't scale 
X_train_pre_ss=X_train.drop(columns=['id', 'pid'])
X_train_ids=X_train[['id', 'pid']]

data_pre_ss=data.drop(columns=['id', 'pid'])
data_ids=data[['id', 'pid']]

#Instantiating Standard Scaler 
ss=StandardScaler()

#Fitting the scaler 

Z_train_noid=pd.DataFrame(ss.fit_transform(X_train_pre_ss), 
                     columns=X_train_pre_ss.columns, index=X_train_pre_ss.index)
data_noid=pd.DataFrame(ss.transform(data_pre_ss), columns=data_pre_ss.columns, index=data_pre_ss.index)

#Join dataframes back together 
Z_train=pd.concat([X_train_ids, Z_train_noid], axis=1)
Z_test_data=pd.concat([data_ids, data_noid], axis=1)

Z_train.head(2)

Z_test_data.head(2)

#### Dropping VIF columns from both datasets 

#Dropping columns in both datasets from the previous VIF
Z_train.drop(columns=['ms_sub_class_20', 'exterior_2nd_VinylSd', 'garage_type_Attchd',
                         'house_style_1Story','roof_style_Gable', 'exterior_2nd_CmentBd',
                         'exterior_1st_MetalSd', 'ms_zoning_RL','house_style_2Story','house_style_SLvl',
                         'mas_vnr_type_None','bldg_type_2fmCon','pool_qc','exterior_1st_VinylSd',
                         'year_built', 'neighborhood_NAmes', 'exterior_1st_HdBoard',
                         'exterior_2nd_Wd Sdng', 'ms_sub_class_60', 'foundation_PConc',
                         'exterior_2nd_Brk Cmn','ms_sub_class_75', 'neighborhood_Somerst',
                         'bsmt_qual', 'fireplace_qu'], inplace=True)

#Dropping columns in both datasets from the previous VIF
Z_test_data.drop(columns=['ms_sub_class_20', 'exterior_2nd_VinylSd', 'garage_type_Attchd',
                         'house_style_1Story','roof_style_Gable', 'exterior_2nd_CmentBd',
                         'exterior_1st_MetalSd', 'ms_zoning_RL','house_style_2Story','house_style_SLvl',
                         'mas_vnr_type_None','bldg_type_2fmCon','pool_qc','exterior_1st_VinylSd',
                         'year_built', 'neighborhood_NAmes', 'exterior_1st_HdBoard',
                         'exterior_2nd_Wd Sdng', 'ms_sub_class_60', 'foundation_PConc',
                         'exterior_2nd_Brk Cmn','ms_sub_class_75', 'neighborhood_Somerst',
                         'bsmt_qual', 'fireplace_qu'], inplace=True)

Z_train.shape

Z_test_data.shape

#Saving dataframes for model 
Z_train.to_csv('Z_train_final.csv')
Z_test_data.to_csv('clean_test.csv')
