==> project-2/Code/part6_test_model_kaggle.ipynb <==
# Project 2: Features, Galore: A Linear Regression Analysis on Home Features Predicting Sale Price
---

## Test Model and Kaggle Submission 

This section includes running the test data through the optimal Linear Regression model. The model used is the regular Linear Regression with all non multi-collinear features. Predicted home sale price values are determined for submission in the Ames Housing Kaggle Competition. 

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.linear_model import LinearRegression, Lasso, Ridge
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.metrics import mean_squared_error
from statsmodels.stats.outliers_influence import variance_inflation_factor

#Read in files 
Z_train=pd.read_csv('../datasets/z_train_final.csv')
Z_test_data=pd.read_csv('../datasets/clean_test.csv')
y_train=pd.read_csv('../datasets/y_train_cleaned.csv')

#Debugging code 
Z_test_data = Z_test_data.replace((np.inf, -np.inf, np.nan), 0).reset_index(drop=True)

Z_train.head()

Z_train=Z_train.set_index('index')

Z_train.head()

Z_train.shape

y_train.shape

y_train.head()

y_train.rename(columns={"Unnamed: 0": "index"}, inplace=True)
y_train=y_train.set_index('index')

y_train.head(2)

Z_test_data.head()

Z_test_data.rename(columns={"Unnamed: 0": "index"}, inplace=True)
Z_test_data=Z_test_data.set_index('index')

Z_test_data.head()

#Renaming unnamed 0 and setting it to the index
Z_test_data.rename(columns={"Unnamed: 0": "index"}, inplace=True)

Z_test_data.head(2)

Z_test_data.head()

#Dropping Utilities because scaled stange 
Z_train.drop(columns='utilities', inplace=True)
Z_test_data.drop(columns='utilities', inplace=True)

Z_test_data_id=Z_test_data['id']

Z_test_data_id.head(2)

#Dropping id and PID for the model 
Z_train=Z_train.drop(columns=['id', 'pid'])
Z_test_data=Z_test_data.drop(columns=['id', 'pid'])

#Instantiate Best Model 
lr=LinearRegression()

y_train.mean()

#fit best model 
lr.fit(Z_train, y_train)

#Train score
lr.score(Z_train, y_train)

#Test data predictions 
lr.predict(Z_test_data)[:10]

#Creating an empty dataframe 
kaggle=pd.DataFrame()

#Making preds into a Dataframe 
kaggle['Id']=Z_test_data_id

#Creating a Sale Price column 
kaggle['SalePrice']=lr.predict(Z_test_data)

#Check 
kaggle.head()

#Check 
kaggle.tail()

#### The below code is not best practice for this assignment but necessary due to issues with kaggle submission

sorted_df = kaggle.sort_values(by=['Id'], ascending=True)

sorted_df = sorted_df.sort_index(ascending=True)

sorted_df=sorted_df.reset_index().drop(columns='Id')
sorted_df

kaggle=sorted_df.rename(columns={'index':'Id', 'SalePrice':'sample_soln'})

kaggle=kaggle[:260]

kaggle

#Saving kaggle as a csv
kaggle.to_csv('submission_lr.csv', index=False)
